Add jwt authentication

+++ app/routers/resources.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, Depends,HTTPException, Request, Form
 from app.database import *
 from app.schemas import *
 from  app.models import*
+from app.auth import *
 from fastapi.templating import Jinja2Templates
 from pathlib import Path
 from typing import List
@@ -17,9 +18,9 @@ def resources_page(request: Request):
     return templates.TemplateResponse("add_resource.html", {"request": request})
 
 @router.post("/resources", response_model= List[ResourceResponse])
-def resource_create(resources: List[ResourceCreate], db= Depends(get_db)):
+def resource_create(resources: List[ResourceCreate], email: str = Depends(get_current_user_email), db= Depends(get_db)):
     try:
-        db_resource= [Resource(**resource.model_dump()) for resource in resources]
+        db_resource= [Resource(**resource.model_dump(exclude={"owner_id"}), owner_id=email) for resource in resources]
         db.add_all(db_resource)
         db.commit()
         return db_resource
@@ -35,10 +36,11 @@ def bulk_update(
     tags: list[str] = Form(...),
     description: list[str] = Form(...),
     id: list[int] = Form(...),
+    email: str = Depends(get_current_user_email),
     db= Depends(get_db)
 ):
     for i, t, u, tg, d in zip(id, title, url, tags, description):
-        res = db.query(Resource).filter(Resource.id == i).first()
+        res = db.query(Resource).filter(Resource.id == i, Resource.owner_id == email).first()
         res.title = t
         res.url = u
         res.tags = tg
@@ -49,14 +51,14 @@ def bulk_update(
     return RedirectResponse(url="/dashboard", status_code=303)
 
 @router.delete("/resources/{id}")
-def resource_delete(id:int, db= Depends(get_db)):
+def resource_delete(id:int, email: str = Depends(get_current_user_email), db= Depends(get_db)):
     
-    db.query(Resource).filter(Resource.id== id).delete()
+    db.query(Resource).filter(Resource.id== id, Resource.owner_id == email).delete()
     db.commit()
 
 @router.get("/dashboard")
-def read_items(request: Request, db = Depends(get_db)):
-    resources= db.query(Resource).all()
+def read_items(request: Request, email: str = Depends(get_current_user_email), db = Depends(get_db)):
+    resources= db.query(Resource).filter(Resource.owner_id == email).all()
     return templates.TemplateResponse("dashboard.html", {"request": request, "resources": resources})
